fix(frontend): address review findings for admin rate-limit management
- Remove useCallback from RateLimitModal (violates project convention);
  inline fetch logic into useEffect body
- Extract useRateLimitManager hook from RateLimitManager.tsx to stay
  under ~200 line file-length guideline
- Compose RateLimitDisplay inside RateLimitModal to eliminate duplicated
  reset UI logic (select/button/confirmation/nothingToReset)
- Add Math.max(0, ...) guard in UsageBar to prevent negative width
