+++ autogpt_platform/frontend/src/app/(platform)/admin/spending/components/RateLimitModal.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { useState, useEffect, useCallback } from "react";
+import { useState, useEffect } from "react";
 import { Button } from "@/components/atoms/Button/Button";
 import {
   Dialog,
@@ -16,7 +16,7 @@ import {
   postV2ResetUserRateLimitUsage,
 } from "@/app/api/__generated__/endpoints/admin/admin";
 import { Gauge } from "@phosphor-icons/react";
-import { UsageBar } from "../../components/UsageBar";
+import { RateLimitDisplay } from "../../rate-limits/components/RateLimitDisplay";
 
 export function RateLimitModal({
   userId,
@@ -30,48 +30,40 @@ export function RateLimitModal({
   const [isLoading, setIsLoading] = useState(false);
   const [rateLimitData, setRateLimitData] =
     useState<UserRateLimitResponse | null>(null);
-  const [resetWeekly, setResetWeekly] = useState(false);
-  const [isResetting, setIsResetting] = useState(false);
-
-  const fetchRateLimit = useCallback(async () => {
-    setIsLoading(true);
-    try {
-      const response = await getV2GetUserRateLimit({ user_id: userId });
-      if (response.status !== 200) {
-        throw new Error("Failed to fetch rate limit");
-      }
-      setRateLimitData(response.data);
-    } catch (error) {
-      console.error("Error fetching rate limit:", error);
-      toast({
-        title: "Error",
-        description: "Failed to fetch user rate limit.",
-        variant: "destructive",
-      });
-      setRateLimitData(null);
-    } finally {
-      setIsLoading(false);
-    }
-  }, [userId, toast]);
 
   useEffect(() => {
-    if (open) {
-      fetchRateLimit();
-    } else {
+    if (!open) {
       setRateLimitData(null);
-      setResetWeekly(false);
+      return;
     }
-  }, [open, fetchRateLimit]);
 
-  async function handleReset() {
+    async function fetchRateLimit() {
+      setIsLoading(true);
+      try {
+        const response = await getV2GetUserRateLimit({ user_id: userId });
+        if (response.status !== 200) {
+          throw new Error("Failed to fetch rate limit");
+        }
+        setRateLimitData(response.data);
+      } catch (error) {
+        console.error("Error fetching rate limit:", error);
+        toast({
+          title: "Error",
+          description: "Failed to fetch user rate limit.",
+          variant: "destructive",
+        });
+        setRateLimitData(null);
+      } finally {
+        setIsLoading(false);
+      }
+    }
+
+    fetchRateLimit();
+  }, [open, userId]);
+
+  async function handleReset(resetWeekly: boolean) {
     if (!rateLimitData) return;
 
-    const msg = resetWeekly
-      ? "Reset both daily and weekly usage counters to zero?"
-      : "Reset daily usage counter to zero?";
-    if (!window.confirm(msg)) return;
-
-    setIsResetting(true);
     try {
       const response = await postV2ResetUserRateLimitUsage({
         user_id: rateLimitData.user_id,
@@ -94,17 +86,8 @@ export function RateLimitModal({
         description: "Failed to reset rate limit usage.",
         variant: "destructive",
       });
-    } finally {
-      setIsResetting(false);
     }
   }
-
-  const nothingToReset = rateLimitData
-    ? resetWeekly
-      ? rateLimitData.daily_tokens_used === 0 &&
-        rateLimitData.weekly_tokens_used === 0
-      : rateLimitData.daily_tokens_used === 0
-    : true;
 
   return (
     <>
@@ -136,48 +119,7 @@ export function RateLimitModal({
           )}
 
           {!isLoading && rateLimitData && (
-            <div className="space-y-6">
-              <div className="grid grid-cols-2 gap-6">
-                <div className="space-y-2">
-                  <h3 className="text-sm font-medium text-gray-700">
-                    Daily Usage
-                  </h3>
-                  <UsageBar
-                    used={rateLimitData.daily_tokens_used}
-                    limit={rateLimitData.daily_token_limit}
-                  />
-                </div>
-                <div className="space-y-2">
-                  <h3 className="text-sm font-medium text-gray-700">
-                    Weekly Usage
-                  </h3>
-                  <UsageBar
-                    used={rateLimitData.weekly_tokens_used}
-                    limit={rateLimitData.weekly_token_limit}
-                  />
-                </div>
-              </div>
-
-              <div className="flex items-center gap-3 border-t pt-4">
-                <select
-                  aria-label="Reset scope"
-                  value={resetWeekly ? "both" : "daily"}
-                  onChange={(e) => setResetWeekly(e.target.value === "both")}
-                  className="rounded-md border bg-white px-3 py-1.5 text-sm"
-                  disabled={isResetting}
-                >
-                  <option value="daily">Reset daily only</option>
-                  <option value="both">Reset daily + weekly</option>
-                </select>
-                <Button
-                  variant="outline"
-                  onClick={handleReset}
-                  disabled={isResetting || nothingToReset}
-                >
-                  {isResetting ? "Resetting..." : "Reset Usage"}
-                </Button>
-              </div>
-            </div>
+            <RateLimitDisplay data={rateLimitData} onReset={handleReset} />
           )}
 
           {!isLoading && !rateLimitData && (
